use common unfold outputs and per-fw concate outputs
