+++ model_compression_toolkit/core/common/hessian/trace_hessian_request.py
@@ -66,6 +66,7 @@ class TraceHessianRequest:
         self.mode = mode  # w.r.t activations or weights
         self.granularity = granularity  # per element, per layer, per channel
         self.target_node = target_node # TODO: extend it list of nodes
+
     def __eq__(self, other):
         # Checks if the other object is an instance of TraceHessianRequest
         # and then checks if all attributes are equal.
