+++ model_compression_toolkit/core/keras/hessian/trace_hessian_calculator_keras.py
@@ -56,144 +56,15 @@ class TraceHessianCalculatorKeras(TraceHessianCalculator):
                                                           trace_hessian_request=trace_hessian_request,
                                                           num_iterations_for_approximation=num_iterations_for_approximation)
 
-    def _get_model_outputs_for_single_image(self,
-                                            output_list: List[str],
-                                            gradient_tape: tf.GradientTape) -> Tuple[List[tf.Tensor], List[tf.Tensor]]:
-        """
-        Computes the model's output according to the given graph representation on the given input,
-        while recording necessary intermediate tensors for gradients computation.
+    def _concat_outputs(self, outputs):
+        unfold_outputs = self._unfold_outputs(outputs)
+        r_outputs = [tf.reshape(output, shape=[output.shape[0], -1]) for output in unfold_outputs]
 
-        Args:
-            output_list: List of nodes that considered as model's output for the purpose of gradients computation.
-            gradient_tape: A GradientTape object for recording necessary info for computing gradients.
+        # Ensure all outputs have the same shape for concatenation
+        concat_axis_dim = [o.shape[0] for o in r_outputs]
+        if not all(d == concat_axis_dim[0] for d in concat_axis_dim):
+            Logger.critical(
+                "Can't concat model's outputs for gradients calculation since the shape of the first axis "  # pragma: no cover  
+                "is not equal in all outputs.")
 
-        Returns: A list of output tensors and a list of activation tensors of all interest points.
-
-        """
-        model_input_tensors = {inode: self.fw_impl.to_tensor(self.input_images[i]) for i, inode in
-                               enumerate(self.graph.get_inputs())}
-
-        node_to_output_tensors_dict = dict()
-
-        # Build an OperationHandler to handle conversions from graph nodes to Keras operators.
-        oh = OperationHandler(self.graph)
-        input_nodes_to_input_tensors = {inode: tf.convert_to_tensor(model_input_tensors[inode]) for
-                                        inode in self.graph.get_inputs()}  # Cast numpy array to tf.Tensor
-
-        interest_points_tensors = []
-        output_tensors = []
-        for n in oh.node_sort:
-            # Build a dictionary from node to its output tensors, by applying the layers sequentially.
-            op_func = oh.get_node_op_function(n)  # Get node operation function
-
-            input_tensors = self._build_input_tensors_list(n,
-                                                           self.graph,
-                                                           node_to_output_tensors_dict)  # Fetch Node inputs
-
-            out_tensors_of_n = self._run_operation(n,  # Run node operation and fetch outputs
-                                                   input_tensors,
-                                                   op_func,
-                                                   input_nodes_to_input_tensors)
-
-            # Gradients can be computed only on float32 tensors
-            if isinstance(out_tensors_of_n, list):
-                for i, t in enumerate(out_tensors_of_n):
-                    out_tensors_of_n[i] = tf.dtypes.cast(t, tf.float32)
-            else:
-                out_tensors_of_n = tf.dtypes.cast(out_tensors_of_n, tf.float32)
-
-            if n.name==self.hessian_request.target_node.name:
-                # Recording the relevant feature maps onto the gradient tape
-                gradient_tape.watch(out_tensors_of_n)
-                interest_points_tensors.append(out_tensors_of_n)
-            if n in output_list:
-                output_tensors.append(out_tensors_of_n)
-
-            if isinstance(out_tensors_of_n, list):
-                node_to_output_tensors_dict.update({n: out_tensors_of_n})
-            else:
-                node_to_output_tensors_dict.update({n: [out_tensors_of_n]})
-
-        return output_tensors, interest_points_tensors
-
-    def _build_input_tensors_list(self,
-                                  node: BaseNode,
-                                  graph: Graph,
-                                  node_to_output_tensors_dict: Dict[BaseNode, List[TFReference]]) -> List[List[TFReference]]:
-        """
-        Given a node, build a list of input tensors the node gets. The list is built
-        based on the node's incoming edges and previous nodes' output tensors.
-
-        Args:
-            node: Node to build its input tensors list.
-            graph: Graph the node is in.
-            node_to_output_tensors_dict: A dictionary from a node to its output tensors.
-
-        Returns:
-            A list of the node's input tensors.
-        """
-
-        input_tensors = []
-        # Go over a sorted list of the node's incoming edges, and for each source node get its output tensors.
-        # Append them in a result list.
-        for ie in graph.incoming_edges(node, sort_by_attr=EDGE_SINK_INDEX):
-            _input_tensors = [node_to_output_tensors_dict[ie.source_node][ie.source_index]]
-            input_tensors.append(_input_tensors)
-        return input_tensors
-
-    def _run_operation(self,
-                       n: BaseNode,
-                       input_tensors: List[List[TFReference]],
-                       op_func: Layer,
-                       input_nodes_to_input_tensors: Dict[BaseNode, Any]) -> List[TFReference]:
-        """
-        Applying the layer (op_func) to the input tensors (input_tensors).
-
-        Args:
-            n: The corresponding node of the layer it runs.
-            input_tensors: List of references to Keras tensors that are the layer's inputs.
-            op_func: Layer to apply to the input tensors.
-            input_nodes_to_input_tensors: A dictionary from a node to its input tensors.
-
-        Returns:
-            A list of references to Keras tensors. The layer's output tensors after applying the
-            layer to the input tensors.
-        """
-
-        if len(input_tensors) == 0:  # Placeholder handling
-            out_tensors_of_n = input_nodes_to_input_tensors[n]
-        else:
-            input_tensors = [tensor for tensor_list in input_tensors for tensor in tensor_list]  # flat list of lists
-            # Build a functional node using its args
-            if isinstance(n, FunctionalNode):
-                if n.inputs_as_list:  # If the first argument should be a list of tensors:
-                    out_tensors_of_n = op_func(input_tensors, *n.op_call_args, **n.op_call_kwargs)
-                else:  # If the input tensors should not be a list but iterated:
-                    out_tensors_of_n = op_func(*input_tensors, *n.op_call_args, **n.op_call_kwargs)
-            else:
-                # If operator expects a single input tensor, it cannot be a list as it should have a dtype field.
-                if len(input_tensors) == 1:
-                    input_tensors = input_tensors[0]
-                out_tensors_of_n = op_func(input_tensors)
-
-        return out_tensors_of_n
-
-
-
-def _concat_outputs(outputs):
-    unfold_outputs = []
-    for output in outputs:
-        if isinstance(output, List):
-            unfold_outputs += output
-        else:
-            unfold_outputs.append(output)
-    r_outputs = [tf.reshape(output, shape=[output.shape[0], -1]) for output in unfold_outputs]
-    # Ensure all outputs have the same shape for concatenation
-    concat_axis_dim = [o.shape[0] for o in r_outputs]
-    if not all(d == concat_axis_dim[0] for d in concat_axis_dim):
-        Logger.critical(
-            "Can't concat model's outputs for gradients calculation since the shape of the first axis "  # 
-            # pragma: no cover  
-            "is not equal in all outputs.")
-    output = tf.concat(r_outputs, axis=1)
-    return output
+        return tf.concat(r_outputs, axis=1)
